+++ apps/cron/src/tasks/processGemsPayout/deployReferralRewardsContract.ts
@@ -1,7 +1,7 @@
 import { log } from '@charmverse/core/log';
 import { prisma } from '@charmverse/core/prisma-client';
-import { getCurrentSeason, getCurrentSeasonWeekNumber, getDateFromISOWeek } from '@packages/dates/utils';
-import { getTopConnectorOfTheDay } from '@packages/scoutgame/topConnector/getTopConnectors';
+import { getCurrentSeason, getCurrentSeasonWeekNumber } from '@packages/dates/utils';
+import { getReferralsToReward } from '@packages/scoutgame/quests/getReferralsToReward';
 import type { DateTime } from 'luxon';
 import { parseUnits } from 'viem';
 import { optimism } from 'viem/chains';
@@ -9,24 +9,13 @@ import { optimism } from 'viem/chains';
 import { createSablierAirdropContract } from './createSablierAirdropContract';
 import { optimismTokenDecimals, optimismTokenAddress } from './deployNewScoutRewardsContract';
 
-const REFERRAL_CHAMPION_REWARD_AMOUNT = parseUnits('25', optimismTokenDecimals).toString();
-
 export async function deployReferralChampionRewardsContract({ week }: { week: string }) {
-  const referralChampions: { address: string; date: DateTime }[] = [];
   const currentSeason = getCurrentSeason(week);
-  const weekStart = getDateFromISOWeek(week).startOf('week');
 
-  for (let day = 0; day <= 6; day++) {
-    const date = weekStart.plus({ days: day });
-    const topConnector = await getTopConnectorOfTheDay({ date });
+  const recipients = await getReferralsToReward({ week });
 
-    if (topConnector) {
-      referralChampions.push({ address: topConnector.address.toLowerCase(), date });
-    }
-  }
-
-  if (referralChampions.length === 0) {
-    log.info('No top connectors found for the week, skipping referral champion rewards contract deployment', {
+  if (recipients.length === 0) {
+    log.info('No referral reward recipients found for the week, skipping referral rewards contract deployment', {
       week,
       season: currentSeason.start
     });
@@ -39,7 +28,7 @@ export async function deployReferralChampionRewardsContract({ week }: { week: st
     chainId: optimism.id,
     tokenAddress: optimismTokenAddress,
     tokenDecimals: optimismTokenDecimals,
-    recipients: referralChampions.map(({ address }) => ({ address: address as `0x${string}`, amount: 25 })),
+    recipients: recipients.map(({ address, opAmount }) => ({ address: address as `0x${string}`, amount: opAmount })),
     nullAddressAmount: 0.001
   });
 
@@ -65,11 +54,11 @@ export async function deployReferralChampionRewardsContract({ week }: { week: st
       deployTxHash: hash,
       rewardPayouts: {
         createMany: {
-          data: referralChampions.map(({ address, date }) => ({
-            amount: REFERRAL_CHAMPION_REWARD_AMOUNT,
+          data: recipients.map(({ address, opAmount }) => ({
+            amount: parseUnits(opAmount.toString(), optimismTokenDecimals).toString(),
             walletAddress: address,
             meta: {
-              date: date.toJSDate()
+              week
             }
           }))
         }
